2020-12-08 On Friendship
Written as prose as part of a writing exercise with Ross.
Haven't been able to put any words together for quite a
while. Although this was originally a single paragraph,
it breaks neatly into stanzas, and is poetic enough to
be a worthy addition.

Thanks to @rossbernet for the encouragement.